Rename user-profile-form to legacy-user-profile-form
